CRUD for Products

+++ FoodDelivery.Data/Migrations/201809021818212_InitialDatabase.cs
@@ -24,7 +24,7 @@ namespace FoodDelivery.Data.Migrations
                     Id = c.Guid(nullable: false, identity: true),
                     Name = c.String(nullable: false, maxLength: 50),
                     Price = c.Decimal(nullable: false, precision: 18, scale: 2),
-                    Mass = c.Double(nullable: false),
+                    Mass = c.Int(nullable: false),
                     CategoryId = c.Guid(nullable: false),
                 })
                 .PrimaryKey(t => t.Id)
